refactor : shop 엔티티 패키지 분리
